feat(miner-foundation): extract predicted-gate types into gittensory-engine (#2276)
Move the PredictedGateVerdict/PredictedGateInput types and the predictedGateNote/
publicSafeFinding pure helpers out of src/rules/predicted-gate.ts into
packages/gittensory-engine/src/predicted-gate.ts, re-exported from the engine barrel,
so a miner can model its own predicted gate verdict locally with the same shapes the
maintainer gate uses. This is the types-and-pure-helpers-first slice;
buildPredictedGateVerdict itself follows in the keystone (#2283).

src/rules/predicted-gate.ts re-exports the two types and imports the helpers via the
relative source path (../../packages/gittensory-engine/src/predicted-gate), matching
this repo's existing engine-consumption convention (see src/scoring/preview.ts), so
typecheck and test:coverage never depend on the engine's built dist/ that is not
guaranteed present when they run in CI. The two small union shapes the surface needs
(GatePolicyPack, GateCheckConclusion) are mirrored engine-side with src kept canonical,
mirroring the scoring/types.ts convention. publicSafeFinding takes its redaction
function as an argument so the engine stays isolated from src/github/commands'
sanitizePublicComment; src binds the canonical sanitizer. The existing predicted-gate
test suite passes unmodified.
